Gate portable Direct2D across GPU backends

+++ src/ProGPU.Native/src/Direct2D/progpu_native_direct2d_render_target.cpp
@@ -807,7 +807,8 @@ private:
     [[nodiscard]] std::uint32_t primitive_flags() const noexcept
     {
         return antialias_mode_ == antialias_mode::aliased
-            ? PROGPU_NATIVE_PRIMITIVE_FLAG_EDGE_ALIASED
+            ? static_cast<std::uint32_t>(
+                PROGPU_NATIVE_PRIMITIVE_FLAG_EDGE_ALIASED)
             : 0U;
     }
